pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dqflow"
version = "0.1.0"
description = "Lightweight, contract-first data quality engine for modern data pipelines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Quang", email = "nguyenvanquang247@gmail.com" },
]
keywords = [
    "data-quality",
    "data-validation",
    "data-contracts",
    "etl",
    "pandas",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pandas>=1.5.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.scripts]
dq = "dqflow.cli:main"

[project.urls]
Homepage = "https://github.com/quang/dqflow"
Repository = "https://github.com/quang/dqflow"
Issues = "https://github.com/quang/dqflow/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/dqflow"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/dqflow"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
